+++ Sources/Shared/Toolkit/HTTP/HTTPServer.swift
@@ -86,7 +86,7 @@ public extension HTTPServer {
                 let link = publication.linkWithHREF(href),
                 let resource = publication.get(href)
             else {
-                onFailure?(request, .access(.http(notFound)))
+                onFailure?(request, ReadError(notFound))
 
                 return HTTPServerResponse(error: notFound)
             }
@@ -136,7 +136,7 @@ public struct HTTPServerResponse: Sendable {
 
     public init(error: HTTPError) {
         self.init(
-            resource: FailureResource(error: .access(.http(error))),
+            resource: FailureResource(error: ReadError(error)),
             mediaType: nil
         )
     }